Modularised the code and renamed the mem files
