+++ plonky2/benches/keccak.rs
@@ -1,43 +1,20 @@
-use plonky2::{plonk::config::PoseidonGoldilocksConfig, util::serialization::Write};
-use plonky2_circuits::PLONKY2_BENCH_PROPERTIES;
-use plonky2_circuits::bench::{keccak256_prepare, prove, verify};
-use plonky2_u32::gates::arithmetic_u32::{U32GateSerializer, U32GeneratorSerializer};
-use utils::harness::ProvingSystem;
+use clap::Parser;
+use plonky2_circuits::bench::{keccak256_prepare, prove};
 
-const D: usize = 2;
-type C = PoseidonGoldilocksConfig;
+#[derive(Parser, Debug)]
+struct Args {
+    /// Input size parameter
+    #[arg(long)]
+    input_size: usize,
+}
 
-utils::define_benchmark_harness!(
-    BenchTarget::Keccak,
-    ProvingSystem::Plonky2,
-    None,
-    "keccak_no_lookup_mem",
-    PLONKY2_BENCH_PROPERTIES,
-    keccak256_prepare,
-    |(_, _, n_gates)| *n_gates,
-    |(circuit_data, pw, _)| { prove(circuit_data, pw.clone()) },
-    |(circuit_data, _pw, _), proof| {
-        let verifier_data = circuit_data.verifier_data();
-        verify(&verifier_data, proof.clone());
-    },
-    |(circuit_data, _pw, _)| {
-        let gate_serializer = U32GateSerializer;
-        let common_data_size = circuit_data
-            .common
-            .to_bytes(&gate_serializer)
-            .unwrap()
-            .len();
-        let generator_serializer = U32GeneratorSerializer::<C, D>::default();
-        let prover_data_size = circuit_data
-            .prover_only
-            .to_bytes(&generator_serializer, &circuit_data.common)
-            .unwrap()
-            .len();
-        prover_data_size + common_data_size
-    },
-    |proof| {
-        let mut buffer = Vec::new();
-        buffer.write_proof(&proof.proof).unwrap();
-        buffer.len()
-    }
-);
+fn main() {
+    let args = Args::parse();
+
+    keccak_mem(args.input_size);
+}
+
+fn keccak_mem(input_size: usize) {
+    let (data, pw, _) = keccak256_prepare(input_size);
+    let _proof = prove(&data, pw);
+}
