+++ plonky2/src/bin/keccak_mem.rs
